Track object expiration with RepositoryTracker
Replace using `HasHashAndUri` with plain `RepositoryEntry` and provide
overloaded `from` for conversion. Use `Stream` for updates to support low-cost
conversion.

Expiration is optional on `RepositoryEntry`. Not because we build open-ended
certificates, but because for RPKI core the expiration is unknown (neither
expiration nor creation is available in the API). Treat these objects as
open-ended, since they are not monitored for expiration anyway.