Reduce memory usage by interning URI and SHA-256 strings
Another option is to use `-XX:+UseStringDeduplication` and let the GC
do this, but this requires support from the GC algorithm (G1 supports
it, not sure about the others) and only deduplicates the underlying
`byte[]` array, not the `String` instance itself.

Explicit `intern` always works and halves the number of instances when
running against a single prepdev instance.