Fix integration tests using outdated RRDP data
Integration tests were failing because the fixed RRDP files are becoming
outdated causing assertion boundaries to fail. Remove the dependency on these
outdated test files.

Replace the fixed files with live RRDP data for integration tests and cache it
for the remainder of the day. Adjust the asserted boundaries to better deal with
changes in the data.

The rationale for using live data is that fixed data will fail the build from
time to time. Especially having this cached in an external dependency makes it
hard to update. Rely on live data instead and allow skipping integration tests
when the data cannot be fetched.